Extract out chat component and add modal view
Extracts chat component (incl. logic). Adds a modal component and
improves the sidebar slightly - darkens rest of the page. Also improved
visual customization.
