+++ examples/nextjs/.gitignore
@@ -28,6 +28,7 @@ yarn-error.log*
 
 # local env files
 .env*.local
+.env.*
 
 # vercel
 .vercel
