+++ examples/nextjs/src/app/page.tsx
@@ -2,9 +2,15 @@
 import { SuperflowsButton } from "@superflows/chat-ui-react";
 import { Cog6ToothIcon } from "@heroicons/react/24/outline";
 
+if (!process.env.NEXT_PUBLIC_SUPERFLOWS_API_KEY) {
+  throw new Error(
+    "You must provide a Superflows API key in the environment variables as NEXT_PUBLIC_SUPERFLOWS_API_KEY",
+  );
+}
+
 export default function Home() {
   return (
-    <main className="flex min-h-screen flex-col items-center justify-between p-12 pt-16">
+    <main className="flex min-h-screen flex-col items-center justify-between p-12 pt-16 pb-2">
       <div className="z-10 w-full max-w-5xl items-center justify-between font-mono text-sm lg:flex">
         <div className="z-10 fixed left-0 inset-y-0 w-60 px-6 py-10 bg-gray-600 flex flex-col ">
           <div className="flex flex-row gap-x-4">
@@ -13,17 +19,17 @@ export default function Home() {
             </div>
             <div className="text-gray-200 text-2xl font-sans">Acme CRM</div>
           </div>
-          <div className="mt-16 flex flex-col animate-pulse">
+          <div className="mt-8 flex flex-col animate-pulse">
             {[1, 2, 3, 4, 5, 6, 7, 8, 9].map((n) => (
-              <>
+              <div key={n}>
                 {n !== 1 && <div className="w-full h-px bg-gray-500" />}
                 <div className="flex flex-row my-4">
                   <div className="h-6 bg-gray-400 rounded-full w-full" />
                 </div>
-              </>
+              </div>
             ))}
           </div>
-          <div className="pt-24 animate-pulse flex space-x-4">
+          <div className="bottom-4 inset-x-6 absolute animate-pulse flex space-x-4">
             <div className="rounded-full bg-gray-200 h-10 w-10" />
             <div className="flex-1 space-y-6 py-1">
               <div className="h-2 bg-gray-200 rounded" />
@@ -43,21 +49,23 @@ export default function Home() {
           }
         >
           <SuperflowsButton
-            superflowsApiKey={process.env.NEXT_PUBLIC_SUPERFLOWS_API_KEY}
-            // superflowsUrl={process.env.NEXT_PUBLIC_SUPERFLOWS_API_URL}
+            superflowsApiKey={process.env.NEXT_PUBLIC_SUPERFLOWS_API_KEY ?? ""}
+            userDescription={process.env.NEXT_PUBLIC_USER_DESCRIPTION}
+            userApiKey={process.env.NEXT_PUBLIC_USER_API_KEY}
             AIname={"Superflows Assistant"}
             styling={{
-              slideoverSide: "right",
+              type: "modal",
+              // modalClasses: "bg-gray-50",
+              // slideoverSide: "right",
               // headerTextColor: "#146ef5",
               // buttonColor: "#146ef5",
             }}
-            mockApiResponses={true}
-            suggestions={[
-              "When am I meeting Alex from Alphabet Inc?",
-              "Who was the top-performing sales rep last quarter?",
-              "What's the most valuable open opportunity?",
-            ]}
+            mockApiResponses={false}
+            suggestions={process.env.NEXT_PUBLIC_SUPERFLOWS_SUGGESTIONS?.split(
+              ",",
+            )}
           />
+
           {[1, 2, 3].map((n) => (
             <div key={n} className="m-1 h-5 w-5 bg-gray-500 rounded-full" />
           ))}
@@ -66,7 +74,9 @@ export default function Home() {
         <div className="ml-24 w-full h-full flex flex-col rounded-md">
           <div className="text-2xl text-gray-600 mb-6 w-44 rounded-full h-7 animate-pulse bg-gray-400" />
           {[
-            1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+            // 1, 2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13, 14, 15, 16, 17, 18, 19,
+            1,
+            2, 3, 4, 5, 6, 7, 8, 9, 10, 11, 12, 13,
           ].map((n) => (
             <div
               key={n}
@@ -81,6 +91,37 @@ export default function Home() {
           ))}
         </div>
       </div>
+      {/*<img className={"absolute inset-0"} src={"/highlevel-bg.png"} />*/}
+      {/*<div className="absolute top-1.5 right-[10.5rem]">*/}
+      {/*  <SuperflowsButton*/}
+      {/*    superflowsApiKey={process.env.NEXT_PUBLIC_SUPERFLOWS_API_KEY ?? ""}*/}
+      {/*    // superflowsUrl={process.env.NEXT_PUBLIC_SUPERFLOWS_API_URL}*/}
+      {/*    userDescription={process.env.NEXT_PUBLIC_USER_DESCRIPTION}*/}
+      {/*    userApiKey={process.env.NEXT_PUBLIC_USER_API_KEY}*/}
+      {/*    AIname={"Superflows Assistant"}*/}
+      {/*    styling={{*/}
+      {/*      type: "modal",*/}
+      {/*      modalClasses: "max-w-6xl bg-gray-50 min-h-[30rem]",*/}
+      {/*      solidIcon: true,*/}
+      {/*      buttonColor: "white",*/}
+      {/*      // slideoverSide: "right",*/}
+      {/*      // headerTextColor: "#146ef5",*/}
+      {/*      // buttonColor: "#146ef5",*/}
+      {/*    }}*/}
+      {/*    // devMode={true}*/}
+      {/*    // mockApiResponses={false}*/}
+      {/*    // suggestions={[*/}
+      {/*    //   "When am I meeting Alex from Alphabet Inc?",*/}
+      {/*    //   "Who was the top-performing sales rep last quarter?",*/}
+      {/*    //   "What's the most valuable open opportunity?",*/}
+      {/*    // ]}*/}
+      {/*    suggestions={process.env.NEXT_PUBLIC_SUPERFLOWS_SUGGESTIONS?.split(*/}
+      {/*      ",",*/}
+      {/*    )}*/}
+      {/*    // TODO: THIS DOESN'T WORK!*/}
+      {/*    buttonStyling="p-1 bg-amber-500 text-black"*/}
+      {/*  />*/}
+      {/*</div>*/}
     </main>
   );
 }
